refactor: Change port types from u32 to u16 for SSH configurations
- update port fields in DataSourceConfig and SSHDataSource to u16
- change DEFAULT_SSH_PORT constant to u16
- improve port validation in with_port method to handle invalid port 0

+++ src/config/types.rs
@@ -9,7 +9,7 @@
  * File Created: 2025-11-17 15:55:16
  *
  * Modified By: mingcheng <mingcheng@apache.org>
- * Last Modified: 2025-11-20 14:19:54
+ * Last Modified: 2025-11-20 14:47:28
  */
 
 use serde::{Deserialize, Serialize};
@@ -85,7 +85,7 @@ pub enum DataSourceConfig {
         host: String,
         user: String,
         #[serde(skip_serializing_if = "Option::is_none")]
-        port: Option<u32>,
+        port: Option<u16>,
         #[serde(skip_serializing_if = "Option::is_none")]
         keyfile: Option<String>,
         #[serde(default = "default_command")]
